requirements.txt
pipecat-ai[daily,cartesia,playht,openai,deepgram,google,silero,anthropic,mcp]
fastapi==0.115.6
uvicorn
python-dotenv
twilio
python-multipart